Create Square Difference.java
Some of the Arithmetic
Now let's try a similar thing in Python. You have a skeleton which reads in two integers from the user, and stores them in variables a and b.

Your task is to write some code that calculates the difference of the squares (a squared minus b squared) and prints it out.

It's not part of the exercise, but you may want to think a little about the differences between the previous Java program and this Python program, especially concerning the handling of the types of the data and the variables (here we only have ints).