Create Input Challenge.java
A Little Input Challenge
So we've had a few programs now that read something in and print something out. We're going to do the same again, but this time we'll do it from "scratch".

The skeleton will just be a class with an empty main method. Your task is to add some code so that it reads in a String and a float, concatenates the float onto the String, then prints out the result. You may not have been explicitly taught every component of what you need to do, but you have seen them all.

Try doing this without looking back at the previous code. You may find the auto complete feature handy in exploring what's possible, or you may want to dive into the API specification. There's also a couple of hints in the skeleton.

If you get stuck, ask a question before just looking back at the code - the copy-paste approach to programming is useful at times, but it doesn't teach you anything if you don't already understand what you're copying, it should only be used when know what to do already.