Create Input Challenge 2.java
An Even Smaller Input Challenge
This time we'll look at doing a similar thing in Python. Again you will (hopefully) notice that this sort of simple task is what Python is built to do, so it will feel a lot more lightweight (if you've programmed before though, you may want to think about where and why this kind of approach changes from an advantage to a disadvantage).

The task here is to prompt the user with the phrase Enter a number: , read in a number, then print That number is even. if it is even, and That number is odd. if it is odd.