Create Debug practice.java
A Little Debugging
We start with a familiar looking Skeleton, the code reads in a String and prints it out with " - Checked!" appened (useful, huh?). However the programmer before you was a bit lazy and submitted buggy code to the repo.

Your task is to find and fix the bugs. Try not to just rewrite the code completely, but look at the error messages and fix the bugs themselves. There are 4 mistakes in the code, one of which is a one you won't have come across before - you may want to ask about that one when you get there!