Create Assessment task 1.java
The Full Magilla (Assessment)
This is the assessment task for this lab. It is due by 9:00AM, Monday 28/3/2022.

For this task you will manipulate multiple arrays, of different types using loops.

The program will build a small database of student names and marks, then print the average mark, the student with the highest mark, and the student with the lowest mark.

In order:

The program will ask the user "How many students are there? ", then read in a positive integer 
𝑛
n (you don't have to deal with bad input).

The program will then read in 
𝑛
n student names with prompts "Enter name of student 
𝑖
i: " and "Enter mark 
𝑖
i: ", where 
𝑖
i is replaced by a number from 1 to 
𝑛
n, in order.

The program will then print "The average mark is: " and print the average mark.

The program will then print "The best student is " and print the student's name.

The program will then print "The worst student is " and print the student's name.

The best and worst student are determined by the highest and lowest mark respectively. The best and worst mark will be unique.

An example with both input and output would be:

How many students are there? 4
Enter name of student 1: Luke
Enter mark 1: 50
Enter name of student 2: Cat
Enter mark 2: 99
Enter name of student 3: Nick
Enter mark 3: 98
Enter name of student 4: Mazza
Enter mark 4: 98
The average mark is: 86.25
The best student is Cat
The worst student is Luke