Create Templated Functions
Templated Functions
We revisit some functions from Week 2 tutorial and add templates to them.  Though templates look a bit scary, their basic usage is actually pretty easy.

Swap function:

template <typename T> void swap(T& a, T& b);
Now we can swap two objects of any type.  The code looks exactly as in week 2 with all the ints  replaced by Ts.

Square function:

template <typename T> T square(const T& a);
Compared to Week 2, now we pass by const reference since this could be applied to a type that is expensive to copy.

Print the elements in the interval between two iterators:

template <typename Iter>
void print(const Iter& begin, const Iter& end);
This function should print out the elements starting from the element pointed to by the iterator begin and going up to but not including the element pointed to by end.  Test out your function on both a std::vector and a std::list!

This kind of interval where we include the first element but exclude the last one is called a half-open interval.  This idiom is very common in standard library algorithms: many algorithms have two iterators as parameters and work on the half-closed interval defined by starting at the first parameter and going up to but not including the second one.

This example shows you how the standard library is able to avoid writing a separate algorithm for each container type.  We can use a template for the iterator type and then write the body of the function using the common functionality that all iterators share.

Calling a templated function:

You can't go wrong with explicitly putting the type with which you want to instantiate the templated function.  For example

int x = 3;
square<int>(x);
However, the compiler can also deduce the type of the template and you do not have to write it.  The same example works as

int x = 3;
square(x);
Since the compiler knows that x is an int, it can deduce that you want to call square with int being the template type here.